+++ HISTORY.rst
@@ -8,6 +8,9 @@ Unreleased Changes
   accumulation raster and a flow accumulation threshold, setting all pixels
   with accumulation above that threshold to 1 and all other valid pixels to 0.
   https://github.com/natcap/pygeoprocessing/issues/272
+* Adding a new function, ``pygeoprocessing.create_raster_from_bounding_box``,
+  that enables the creation of a new raster from a bounding box.
+  https://github.com/natcap/pygeoprocessing/issues/276
 * Win32 wheels of PyGeoprocessing are no longer created through our GitHub
   Actions workflows and will no longer be produced or distributed as part of
   our release checklist.  For details (and metrics!) see: